+++ src/mainwindow.cpp
@@ -11487,7 +11487,9 @@ void MainWindow::on_noteEditTabWidget_currentChanged(int index) {
 
     // Allow the subfolder of the note to be selected in the subfolder list
     // See: https://github.com/pbek/QOwnNotes/issues/2861
-    setShowNotesFromAllNoteSubFolders(false);
+    if (QSettings().value(QStringLiteral("noteSubfoldersPanelTabsUnsetAllNotesSelection")).toBool()) {
+        setShowNotesFromAllNoteSubFolders(false);
+    }
 
     setCurrentNoteFromNoteId(noteId);
     widget->setLayout(ui->noteEditTabWidgetLayout);